working with JSON format and postman, still no lombok

+++ bank/src/main/java/com/example/demo/services/KredytService.java
@@ -6,7 +6,7 @@ import com.example.demo.models.KredytModel;
 
 @Service
 public class KredytService {
-	public double Calculate(double amount, double percent, double installmentAmount) {
+	public double Calculate(double amount, double percent, int installmentAmount) {
 		double m = 1 - 1 / Math.pow(1.0 + percent / 12, installmentAmount);
 		double installment = amount * (percent / 12) / m;
 		return installment;
